examples/basic/requirements.txt
alabaster==0.7.6
alembic==0.7.7
amqp==1.4.6
aniso8601==1.1.0
ansi==0.1.3
anyjson==0.3.3
argh==0.26.1
atomicwrites==1.0.0
Babel==2.3.4
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.0
billiard==3.3.0.20
binaryornot==0.4.0
bleach==1.4.1
blessings==1.6
blinker==1.4
bpython==0.15
butterfly==2.0.1
cached-property==1.3.0
-e git+https://github.com/neo1218/cat.git@5d7c66424ed825c1da5e1eeaf0db61e32c61c898#egg=catg
celery==3.1.18
Cerberus==0.8.1
certifi==2015.11.20.1
cffi==1.5.2
characteristic==14.3.0
chardet==2.3.0
charlockholmes==0.0.2
click==6.2
colorama==0.3.3
ConfigArgParse==0.10.0
configobj==5.0.6
configparser==3.3.0.post2
construct==2.5.2
cookiecutter==1.3.0
coverage==4.0
cryptography==1.1.1
cssselect==0.9.1
curtsies==0.2.6
cycler==0.10.0
Cython==0.24
decorator==4.0.4
distribution==1.0.0
Django==1.9
docker-compose==1.6.2
docker-py==0.5.3
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
dominate==2.1.12
douban-client==0.0.6
drawille==0.0.4
dreampie==1.2.1
ecdsa==0.13
enum34==1.1.2
Eve==0.5.3
Events==0.2.1
Fabric==1.11.1
fbone==0.1
fig==1.0.1
flake8==2.5.4
-e git+https://github.com/pallets/flask.git@80b9453bdd9a5e9af9c4d46615b93f4df618e94b#egg=Flask
Flask-Admin==1.2.0
Flask-Babel==0.10.0
Flask-Bootstrap==3.3.5.6
Flask-Cache==0.13.1
Flask-DebugToolbar==0.10.0
Flask-FlatPages==0.6
Flask-HTTPAuth==2.6.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
flask-markdown2==0.0.3
Flask-Migrate==1.5.0
Flask-Misaka==0.3.0
Flask-Moment==0.5.1
Flask-OAuthlib==0.9.2
Flask-OpenID==1.2.4
Flask-PageDown==0.2.1
Flask-PyMongo==0.3.1
Flask-Redis==0.1.0
Flask-RESTful==0.3.5
Flask-Restless==0.17.0
Flask-Script==2.0.5
Flask-SocketIO==2.4
Flask-SQLAlchemy==2.0
Flask-Testing==0.4.2
Flask-Token==1.0
Flask-WhooshAlchemy==0.56
Flask-WTF==0.12
-e git+https://github.com/neo1218/flask-ckeditor.git@71f15f91794107cb0d935f29db881a03d3c76d8b#egg=flaskckeditor
ForgeryPy==0.1
Frozen-Flask==0.12
frozendict==0.6
funcsigs==1.0.2
functools32==3.2.3.post2
future==0.15.2
futures==3.0.5
gevent==1.1.1
gnureadline==6.3.3
greenlet==0.4.9
-e git+https://github.com/neo1218/hacks.git@c813d752e90b73f2a76b7d037d83327943ebebf9#egg=hacks
hjson==2.0.0
hpack==2.0.1
html2text==2015.11.4
html5lib==0.999999
http==0.2
http-prompt==0.4.0
httpie==0.9.2
icalendar==3.9.2
idna==2.0
ipaddress==1.0.16
ipython==3.2.1
itsdangerous==0.24
jedi==0.9.0
jieba==0.38
Jinja2==2.8
jinja2-markdown==0.0.3
jsonschema==2.5.1
Kivy==1.9.1
Kivy-Garden==0.1.4
kombu==3.0.26
libsass==0.11.1
linguist==0.1.1
lxml==3.4.4
Mako==1.0.1
-e git+https://github.com/neo1218/mana.git@6cea7a98e0b527685d0d9d170cbb908dc9ec34b5#egg=mana
mana0==0.10
mana2==0.15
mana3==0.15
mana4==0.15
Markdown==2.6.5
markdown2==2.3.0
markerlib==0.6.0
MarkupSafe==0.23
marshmallow==2.4.2
marshmallow-jsonapi==0.4.2
matplotlib==1.5.1
mccabe==0.4.0
meld3==1.0.2
mercurial==3.8.2
mime==0.0.3
mimerender==0.5.5
misaka==1.0.2
mitmproxy==0.15
mock==2.0.0
mongokit==0.9.1.1
msgpack-python==0.4.7
muxi==0.11
mycli==1.8.0
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
neovim==0.1.8
netlib==0.15.1
nose==1.3.7
numpy==1.11.1
oauthlib==1.0.3
objgraph==3.0.0
paramiko==1.17.0
parsedatetime==2.1
parsimonious==0.6.2
passlib==1.6.5
pathlib==1.0.1
pathtools==0.1.2
pbr==1.10.0
pep8==1.7.0
Pillow==3.0.0
pluggy==0.3.1
powerline-status==2.3.dev9999+git.8e5bc2538b27ffa2706325da79502a664ffb564c
prettytable==0.7.2
prompt-toolkit==1.0.5
psutil==3.2.2
ptpython==0.31
pudb==2015.3
py==1.4.30
py-oauth2==0.0.10
pyaes==1.3.0
pyasn1==0.1.9
pyasn1-modules==0.0.7
PyAudio==0.2.9
pycparser==2.14
pycrypto==2.6.1
pyDAL==15.12
pyflakes==0.9.2
Pygments==2.1
pygments-github-lexers==0.1
pymongo==2.8.1
PyMySQL==0.7.6
pyOpenSSL==0.15.1
pyparsing==2.0.7
pyperclip==1.5.25
pyte==0.4.10
pytest==2.8.2
python-dateutil==2.2
python-editor==1.0
python-engineio==0.9.1
python-mimeparse==0.1.4
python-openid==2.2.5
python-socketio==1.3
pytz==2016.4
pyvim==0.0.16
pyxdg==0.25
PyYAML==3.11
pyzic==1.0
qiniu==7.0.6
queuelib==1.3.0
redis==2.10.3
requests==2.9.1
requests-oauthlib==0.6.1
-e git+https://github.com/neo1218/rest.git@8bf7369aaa9da5cc4a300c625e4d7fea21f52681#egg=rest
ruamel.base==1.0.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.10.12
scanner==0.0.5
Scrapy==1.0.3
selenium==2.47.3
service-identity==14.0.0
setuptools-scm==1.11.0
shadowsocks==2.8.2
-e git+https://github.com/neo1218/ship.git@c74d787f4bad7befd52aee62693d9700fe543dee#egg=ship_cli
simple0==1.2
simplejson==3.8.0
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.0
speaklater==1.3
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
SQLAlchemy==1.0.8
sqlparse==0.1.19
supervisor==3.3.0
svn==0.3.36
texttable==0.8.4
thefuck==3.2
todoman==1.5.3
tornado==4.3
tornado-systemd==1.0.1
tox==2.1.1
tqdm==3.7.1
trollius==2.1
twill==1.8.0
uncompyle2==1.1
UnParse==0.1.3
urwid==1.3.1
vboxapi==1.0
virtualenv==15.0.1
w3lib==1.12.0
watchdog==0.8.3
wcwidth==0.1.6
websocket-client==0.11.0
-e git+https://github.com/gi0baro/weppy.git@18c4348ce914ffe5b6ed8d7257124a461f9266e2#egg=weppy
Werkzeug==0.11.4
whatportis==0.4
whichcraft==0.1.1
Whoosh==2.7.0
WTForms==2.1
wxPython==3.0.2.0
wxPython-common==3.0.2.0
xdot==0.6
